Move overview to dedicated submodule(#45)
