+++ tests/overview/test_dialogs.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+
+
+class TestOverviewDialog(object):
+    """Unittests for the overview dialog."""
+
+    def test_daterange(self, request):
+        """Test that we return the right attribute."""
+        pass
+
+    def test_daterange_emit_signal(self, request):
+        """Test that setting a daterange emit the right signal."""
+        pass
